+++ README_HIDDEN.md
@@ -1,0 +1,2 @@
+## Logs
+Debug and trace logs may contain secrets and are to be treated as if they do
