+++ apps/erudio/backend-api/src/helpers/config.rs
@@ -1,4 +1,6 @@
 use serde::Deserialize;
+use services::s3::{Buckets, Credentials};
+use std::collections::HashMap;
 
 pub const SESSION_COOKIE_NAME: &str = "SessionId";
 // Config file?
@@ -11,6 +13,14 @@ pub struct Config {
 	pub redis_url: String,
 	pub region_id: String,
 	pub api_port: u16,
+	pub buckets: HashMap<Buckets, BucketConfig>,
+}
+
+#[derive(Debug, Clone, Deserialize)]
+pub struct BucketConfig {
+	pub name: String,
+	pub region: String,
+	//pub credentials: Credentials,
 }
 
 #[derive(Debug, Clone, Deserialize)]
