+++ shared-libs/rust/services/src/s3/mod.rs
@@ -1,5 +1,7 @@
-pub use s3::Bucket;
+pub use s3::creds::Credentials;
+use serde::Deserialize;
 
+#[derive(Debug, Clone, Deserialize, Ord, PartialOrd, Eq, PartialEq, Hash)]
 pub enum Buckets {
 	MessageAttachments,
 }
